requirements.txt
--find-links https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113
git+https://github.com/pytorch/botorch.git@c2cb5bc6a2e61b2612ba505865935b681861298d
# git+https://github.com/cornellius-gp/gpytorch.git@307ef9c7b56dee3635dd9e7d06a03bf51f66f6e1
git+https://github.com/samuelstanton/upcycle.git@master
gpytorch==1.6.0
setuptools==60.2.0
wandb==0.10.28
biopython==1.79
omegaconf==2.2.2
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
scikit-learn==1.0.2
pandas==1.3.0
matplotlib==3.5.1
seaborn==0.11.2
cachetools==5.0.0
pymoo==0.5.0
tabulate==0.8.9
python-levenshtein==0.12.2
transformers==4.19.2

# for Nullrecurrent
plotly==5.9.0
h5py==2.10.0
tensorflow==2.9.1
ipynb==0.5.1
numpy==1.21.0

